Proper MSVC support

+++ debug.h
@@ -2,6 +2,7 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stddef.h>
 struct mobile_adapter;
 struct mobile_packet;
 struct mobile_addr;
@@ -13,9 +14,9 @@ struct mobile_adapter_debug {
     char buffer[MOBILE_DEBUG_BUFFER_SIZE];
 };
 
-void mobile_debug_write(struct mobile_adapter *adapter, const char *data, unsigned size);
+void mobile_debug_write(struct mobile_adapter *adapter, const char *data, size_t size);
 void mobile_debug_print(struct mobile_adapter *adapter, const char *fmt, ...);
-void mobile_debug_print_hex(struct mobile_adapter *adapter, const void *data, unsigned size);
+void mobile_debug_print_hex(struct mobile_adapter *adapter, const void *data, size_t size);
 void mobile_debug_print_addr(struct mobile_adapter *adapter, const struct mobile_addr *addr);
 void mobile_debug_endl(struct mobile_adapter *adapter);
 void mobile_debug_init(struct mobile_adapter *adapter);
